Move ContactDetails base class
We will soon have Apply pages that want to use this base class outside of the contact details task, so we move this and the associated Nunjucks partial to a better shared location for use by pages in different tasks
