+++ server/form-pages/apply/accommodation-need/contact-details/probationPractitioner.ts
@@ -1,6 +1,6 @@
 import TasklistPage from '../../../tasklistPage'
 import { Page } from '../../../utils/decorators'
-import ContactDetails from './contactDetails'
+import ContactDetails from '../../contactDetails'
 
 @Page({ name: 'probation-practitioner', bodyProperties: ['name', 'phone', 'email'] })
 export default class ProbationPractitioner extends ContactDetails implements TasklistPage {
