+++ server/views/applications/pages/contact-details/probation-practitioner.njk
@@ -7,6 +7,6 @@
     {{ page.title }}
   </h1>
 
-  {% include "./partials/_contact-details.njk" %}
+  {% include "../partials/_contact-details.njk" %}
 
 {% endblock %}
